requirements.txt.txt
spade==3.3.1
python-dotenv